I2C dispose fix (#3121)

+++ targets/ESP32/_nanoCLR/targetPAL_I2c.cpp
@@ -7,7 +7,7 @@
 #include <esp32_idf.h>
 #include <nanoHAL_v2.h>
 
-uint8_t Esp_I2C_Initialised_Flag[I2C_NUM_MAX] = {
+int16_t Esp_I2C_Initialised_Flag[I2C_NUM_MAX] = {
     0
 #if I2C_NUM_MAX > 1
     ,
